+++ test/fixtures/commonjs_powerassert/fixture.js
@@ -1,0 +1,10 @@
+'use strict';
+
+var assert = require('power-assert');
+
+function add (a, b) {
+    assert(!isNaN(a));
+    assert.equal(typeof b, 'number');
+    assert.ok(!isNaN(b));
+    return a + b;
+}