+++ test/fixtures/commonjs_singlevar/expected.js
@@ -1,0 +1,5 @@
+'use strict';
+var foo = 'FOO', bar = 'BAR';
+function add(a, b) {
+    return a + b;
+}