+++ test/fixtures/commonjs_singlevar/fixture.js
@@ -1,0 +1,12 @@
+'use strict';
+
+var foo = 'FOO',
+    assert = require('assert'),
+    bar = 'BAR';
+
+function add (a, b) {
+    assert(!isNaN(a));
+    assert.equal(typeof b, 'number');
+    assert.ok(!isNaN(b));
+    return a + b;
+}