+++ test/fixtures/func/expected.js
@@ -1,0 +1,4 @@
+'use strict';
+function add(a, b) {
+    return a + b;
+}