+++ test/fixtures/func/fixture.js
@@ -1,0 +1,9 @@
+'use strict';
+
+function add (a, b) {
+    console.assert(typeof a === 'number');
+    assert(!isNaN(a));
+    assert.equal(typeof b, 'number');
+    assert.ok(!isNaN(b));
+    return a + b;
+}