+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [1.1.0](https://github.com/twada/unassert/releases/tag/v1.1.0) (2015-08-11)
+
+
+#### Features
+
+  * remove CommonJS assert declaration ([1c3dc425](https://github.com/twada/unassert/commit/1c3dc425f93f1d8b3790e1ea909a14ff0a6f076f))
+  * remove CommonJS power-assert declaration as well ([5925b38a](https://github.com/twada/unassert/commit/5925b38a351596afab4de2f027fed9dc2ed82602))
+
+
 ## [1.0.0](https://github.com/twada/unassert/releases/tag/v1.0.0) (2015-05-27)